Used dispatch and action creators to dispatch data to the specified reducer and connected the stored data in state to the req component using useSelector hook.
