+++ src/api/index.js
@@ -1,0 +1,7 @@
+import Chance from 'chance';
+
+const chance = Chance();
+
+export const mockData = () => {
+	return chance.name({ middle: true });
+};
