+++ src/components/admin_user_table/AppU.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Navbar from './reusables/Navbar';
 import './index.css';
-import UserDetails from './reusables/UserDetails';
+import UserDetails from './UserDetails';
 
 const AppU = () => {
 	return (
